+++ db.go
@@ -179,5 +179,3 @@ func migrateEpisodes(db *sql.DB) error {
 	}
 	return nil
 }
-
-